Adding code to calculate relative risk indicators
Base code for calculating relative risk indicators as outline in the overleaf rel risk section. Have a couple of questions but a starting point (for me, anyways). I only ran on health-based violations and commented in the values that I derived from the code. Question: I used the population served (not ACSTOTPOP) for each PWSID, though the racial categories were calculated using the ACS (EJscreen) percentages, so an assumption I am making but maybe a wrong one?