Add rel_risk_by_race
Code to calculate relative risk indicators by race category; some interesting heterogeneity in these results. Maybe not the most efficient code but the most time efficient (tried for loops but it was giving me a bunch of errors, so I decided to opt for this version instead, which is not too burdensome).