Update the relative risk indicators
Just updated the values based on the new generated pwsid datasets (namely, filling in zeroes for the NA values for HBV, LCR, and PFAS). The overall risk indicators did not change much but the by_race indicators changed quite a bit. Double check some values because I want to make sure that I did not do voodoo magic on any of these.