Update 00e_hm_boundaries.R
Uses the shape file that I had to pass through ArcGIS to bypass the ZM dimension issue; includes some commented out "validity checks" just in case we need to update with new boundaries and want to verify we're using valid polygons. Removes a PWS in delaware causing the multigeometry issue when running ejfunction.