New code for scatter plots, indicator regression updates, and national map code edits
new code to produce scatter plots that compare demographics across each boundary type. Fixes to the regression code so that the stargazer table is produced with 6 columns. Minor updates to the code to produce national maps of each indicator with respect to the legends, file directories, and censored levels. Lots of new output files for maps, bar charts, and scatterplots